:recycle: Move package.json retrieval to environment module
